Journal: replace the journal throttle with fullness backoff throttle
The existing FileJournal::throttle_(ops|bytes) throttles overlap with
the FileStore op queue throttles.  It doesn't seem important whether
pending ops are waiting on the journal or the backing fs, so the
FileJournal ones are out.  Instead, there is now a throttle which
is taken in queue_transaction and released in _committed_thru
(after sync) which reflects the current fullness of the journal
and gradually delays ops as the journal fills up.  The intention is
to smooth out workloads on small journals.

Signed-off-by: Samuel Just <sjust@redhat.com>